Add basic working code editor
Kinda based off of the ideas here (https://marijnhaverbeke.nl/blog/browser-input-reading.html
), ie. a hidden 'textarea' where input events go to, while the resulting
'code' is rendered separately. This implementation is made much simpler
by the fact that it doesn't need to handle large files, so the entirety
of the document can stay in the textarea.

Seems to work well enough, but I'm not sure it's what I
need--considering we're going to need to be able to mix HTML elements
into the source code. Although I'm also not sure that the pure
contenteditable route is the right path either!