Slight usability tweaks to editor for highlighting
Unfortunately, think this is the wrong path. Turns out it's hard to
control what is added to a textarea, which is important for any sort of
"semantic chunking" of statements that I want to do.

Also implementing any sort of click/drag interface seems like it would
be a special sort of hell.